+++ src/__tests__/functional-tests/form-field-controls-integration.test.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 import { render } from '@testing-library/react';
 
 import FormField, { FormFieldProps } from '../../../lib/components/form-field';
-import { FormFieldValidationControlProps } from '../../../lib/components/internal/context/form-field-context';
 import createWrapper, { ElementWrapper } from '../../../lib/components/test-utils/dom';
+import { FormFieldValidationControlProps } from '../../../types/form-field';
 import { getRequiredPropsForComponent } from '../required-props-for-components';
 import { requireComponent } from '../utils';
