+++ src/autosuggest/interfaces.ts
@@ -6,9 +6,9 @@ import { BaseDropdownHostProps, OptionsFilteringType, OptionsLoadItemsDetail } f
 import { BaseInputProps, InputAutoCorrect, InputClearLabel, InputKeyEvents, InputProps } from '../input/interfaces';
 import { DropdownStatusProps } from '../internal/components/dropdown-status';
 import { OptionDefinition, OptionGroup } from '../internal/components/option/interfaces';
-import { FormFieldValidationControlProps } from '../internal/context/form-field-context';
 import { BaseComponentProps } from '../types/base-component';
 import { NonCancelableEventHandler } from '../types/events';
+import { FormFieldValidationControlProps } from '../types/form-field';
 
 export interface AutosuggestProps
   extends BaseComponentProps,
