+++ src/checkbox/base-checkbox.tsx
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 import React from 'react';
 
-import { FormFieldControlProps } from '../internal/context/form-field-context';
 import { BaseComponentProps } from '../types/base-component';
 import { NonCancelableEventHandler } from '../types/events';
+import { FormFieldControlProps } from '../types/form-field';
 
 export interface BaseCheckboxProps extends BaseComponentProps, FormFieldControlProps {
   /**
