+++ src/contexts/form-field.ts
@@ -1,7 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export {
-  FormFieldValidationControlProps as FormFieldContextValue,
-  useFormFieldContext,
-} from '../internal/context/form-field-context';
+export { useFormFieldContext } from '../internal/context/form-field-context';
+export { FormFieldValidationControlProps as FormFieldContextValue } from '../types/form-field';
