+++ src/date-input/interfaces.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { BaseInputProps, InputProps } from '../input/interfaces';
-import { FormFieldValidationControlProps } from '../internal/context/form-field-context';
 import { DateFormat, DateGranularity, EditableDateFormat } from '../internal/utils/date-time/interfaces';
 import { BaseComponentProps } from '../types/base-component';
+import { FormFieldValidationControlProps } from '../types/form-field';
 
 export interface DateInputProps extends BaseInputProps, FormFieldValidationControlProps, BaseComponentProps {
   /**
