+++ src/form-field/__tests__/form-field-context.test.tsx
@@ -4,11 +4,9 @@ import React from 'react';
 import { render } from '@testing-library/react';
 
 import FormField, { FormFieldProps } from '../../../lib/components/form-field';
-import {
-  FormFieldValidationControlProps,
-  useFormFieldContext,
-} from '../../../lib/components/internal/context/form-field-context';
+import { useFormFieldContext } from '../../../lib/components/internal/context/form-field-context';
 import createWrapper, { ElementWrapper } from '../../../lib/components/test-utils/dom';
+import { FormFieldValidationControlProps } from '../../types/form-field';
 
 import styles from '../../../lib/components/form-field/styles.css.js';
