+++ src/internal/components/autosuggest-input/index.tsx
@@ -20,9 +20,10 @@ import {
 import InternalInput from '../../../input/internal';
 import { BaseComponentProps } from '../../../types/base-component';
 import { BaseKeyDetail, NonCancelableEventHandler } from '../../../types/events';
+import { FormFieldValidationControlProps } from '../../../types/form-field';
 import { getBaseProps } from '../../base-component';
 import { getBreakpointValue } from '../../breakpoints';
-import { FormFieldValidationControlProps, useFormFieldContext } from '../../context/form-field-context';
+import { useFormFieldContext } from '../../context/form-field-context';
 import { fireCancelableEvent, fireNonCancelableEvent } from '../../events';
 import { InternalBaseComponentProps } from '../../hooks/use-base-component';
 import { useIMEComposition } from '../../hooks/use-ime-composition';
