+++ src/radio-group/interfaces.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { RadioButtonProps } from '../internal/components/radio-button/interfaces';
-import { FormFieldControlProps } from '../internal/context/form-field-context';
 import { BaseComponentProps } from '../types/base-component';
 import { NonCancelableEventHandler } from '../types/events';
+import { FormFieldControlProps } from '../types/form-field';
 
 export interface RadioGroupProps extends BaseComponentProps, FormFieldControlProps {
   /**
