+++ src/tiles/interfaces.ts
@@ -3,9 +3,9 @@
 import React from 'react';
 
 import { Breakpoint as _Breakpoint } from '../internal/breakpoints';
-import { FormFieldControlProps } from '../internal/context/form-field-context';
 import { BaseComponentProps } from '../types/base-component';
 import { NonCancelableEventHandler } from '../types/events';
+import { FormFieldControlProps } from '../types/form-field';
 
 export interface TilesProps extends BaseComponentProps, FormFieldControlProps {
   /**
